requirements.txt
requests
mutagen
pycryptodome
